agent/requirements.txt
langchain
langgraph
langchain-community
tavily-python
langchain-core~=0.3.15
langchain_openai
langchain_core
langgraph
pydantic
python-dotenv
json5
copilotkit==0.1.70
langgraph-cli==0.1.71
httpx>=0.27.0
